requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[crytography]
pydantic[email]
python-multipart
psycopg2-binary
prometheus-fastapi-instrumentator

